Bit of refactoring
- Reverted the prisma schema ChatRoomStatus back to default values (was causing issues)
- Moved UserList to chat, renamed files
